+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 
+## [0.2.1] - 2018-23-07
+### Changed
+Ignore late arriving responses when they conflict with requests that have
+been fulfilled more recently.
+
 ## [0.2.0] - 2018-19-07
 ### Changed
 - Remove DOM node element state from unmounting process @Exbia